+++ src/pages/bookinfo/[id].js
@@ -1,6 +1,9 @@
+import SearchBar from "@/components/SearchBar";
+
 export default function Detail({id}) {
   return (
     <div>
+        <SearchBar/>
         <h1>Detail {id}</h1>
     </div>
   );
